Add -v (verbose) option.
Upstream pg_verify_checksums has renamed the -d (debug) option to -v (verbose)
and changed it such that it only outputs every checksummed file. As we have
added additional debug output, we keep the debug option for now but add -v as
well and make it behave identical. The long --verbose option will be added once
support for long options is present.