Fix scanning of tablespaces.
Commit 60fae02 introduced a whitelist of relation filenames to compare files or
directories to. However, it missed that non-default tablespaces have a
PG_<VER>_<CATNO> directory, that was not considered a relation and skipped, so
pg_checksums would not verify or activate checksums in tablespaces.

Fix this by adopting a proposed patch by Michael Paquier that only checks
proper files for being a relation file, not directories. Also adapt TAP tests
inducing corruption in a table in a non-default tablespace.