Port fsync_pgdata() and assorted helper functions from PostgreSQL.
This adds additional code, but allows for dropping the findInitDB() and
syncDataDir() function.  This also means we are longer calling an external
initdb program which turned out to be problematic as it was checking for the
exact minor version it was compiled with thus breaking if PostgreSQL got
updated but pg_checksums was not recompiled after minor releases.