Overhaul short read logic.
If a short read occurs, we now only retry if online.  Also, we skip to the next
block if we still get a short read on re-read and the next file on all other
errors instead of outright erroring out.  A new counter skippedfiles is added
to report those at the end, if any files were skipped due to short reads.  TAP
tests have been adjusted accordingly.