Implement a basic I/O throttling mechanism.
This hooks into the show_progress reporting, which at the moment has a one
second granularity.  If the current rate is more than the desired rate, we
sleep for the required number of seconds to bring the rate down to what was
mandated by the user via the new --max-rate option.

Future enhancement will likely make the time-resolution more fine-grained to
allow for better rate-limiting.