Fine-tune I/O throttling and switch from kB to MB as units.
See whether we should throttle every 50 ms while report the progress every 400
ms (if desired).  Further, convert the max-rate to per-ms and compute the
waiting time in milliseconds instead of seconds in order to increase precision.

While we are at it, report current rate and sizes in MB instead of kB.

Also, only call report_progress_or_throttle() every 100 blocks in order to
lower the overhead of this feature.