Fix checksum failure detection for all-zero or random pageheaders.
Previously, a zeroed-out pageheader would lead to that page getting skipped as
being new, even if other parts of the page have (corrupted) data in it.  Add an
additional check that the whole page shall be zero and report a checksum
failure if that is not the case.

Also, random data in the pageheader would very likely make the page's LSN so
large that it would be considered newer than the checkpoint LSN.  As we are not
connected to the server, we cannot inquire the current insert record pointer to
check against or lock the page, so we (i) advance the checkpoint LSN to the
current pg_control value and (ii) demand that the upper 32 bits of the LSN are
identical between the checkpoint LSN and the page's LSN in order to skip the
page rather than report a checksum failure.

Additional TAP tests are added for both cases as well.