Check PostgreSQL data directory major version.
Previously, running a pg_checksums executable compiled for a particular
PostgreSQL major version against an instance with a different version
could result in the obscure error message `pg_control CRC value is
incorrect'. Instead, adapt the CheckDataVersion() function from
pg_resetwal, which compares PG_VERSION in the data directory to the
compiled-against major version number and errors out if they are
different. The error message in this case is the much more helpful `data
directory is of wrong version', folllowed by the PG_VERSION and
compiled-in major version numbers.