Merge with upstream pg_checksums.
Notably, include an optimization for pg_checksums --enable where checksum is
already set (4c9f50d116461617848601e97dbc6f122b0a6f14). This prevents
pg_checksums to do a rewrite of a block if it has no need to, in the case where
the computed checksum matches with what's already stored in the block read.