Do not consider empty new pages a problem.
If a page is new, we check whether it consists only of zeros and mark it as a
checksum failure if this is not the case. If a page is new and all zero, we
marked this page as skipped for the statistics, however, offline checksum
verifications are considered as failed if there are skipped blocks.

Up to PostgreSQL v15 this was not a problem, but the regression tests started
to fail with v16 as some blocks were marked as skipped. To avoid this, consider
new but empty blocks as scanned but not skipped.