Fix build failure with PostgreSQL 17.
PostgreSQL 17 renamed fsync_pgdata() to sync_pgdata() and added
sync_method as additional argument. To fix building for PostgreSQL 17,
define sync_method as DATA_DIR_SYNC_METHOD_FSYNC and change the function
call to sync_pgdata() for that version. In a follow-up commit, the sync
method will be made configurable.